reception service added

+++ dataspinacademy/src/main/java/com/dataspin/dataspinacademy/projection/CoursePriceInfo.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 public interface CoursePriceInfo {
     Long getId();
 
-    BigDecimal getPrice();
+    Integer getPrice();
 
     LocalDateTime getStartDate();
